get rid of IntegrationTestHelper
