+++ client/src/features/sceneControls/mathItems/__tests__/variable.spec.tsx
@@ -1,11 +1,10 @@
 import { MathItemType as MIT } from "configs";
-import { IntegrationTest, makeItem, screen } from "test_util";
+import { makeItem, renderTestApp, screen, seedDb } from "test_util";
 
 test("left-hand parse errors are indicated on left-hand side", async () => {
   const variable = makeItem(MIT.Variable, { value: "a+ = 123" });
-  const helper = new IntegrationTest();
-  helper.patchMathItemsInFolder([variable]);
-  const { store } = helper.render();
+  const scene = seedDb.withSceneFromItems([variable]);
+  const { store } = await renderTestApp(`/${scene.id}`);
 
   const mathScope = store.getState().mathItems.mathScope();
   expect(mathScope.parseErrors.size).toBe(1);
@@ -17,9 +16,8 @@ test("left-hand parse errors are indicated on left-hand side", async () => {
 
 test("right-hand parse errors are indicated on right-hand side", async () => {
   const variable = makeItem(MIT.Variable, { value: "a = 123 + " });
-  const helper = new IntegrationTest();
-  helper.patchMathItemsInFolder([variable]);
-  const { store } = helper.render();
+  const scene = seedDb.withSceneFromItems([variable]);
+  const { store } = await renderTestApp(`/${scene.id}`);
 
   const mathScope = store.getState().mathItems.mathScope();
   expect(mathScope.parseErrors.size).toBe(1);
